requirements.txt
pydub
PyPDF2
requests
transformers
googletrans
boto3
pillow